Change the data type of the discountPercentage attribute from BigDecimal to float

+++ src/main/java/com/bookingbustickets/bookingbustickets/controller/request/RequestPassengerCategoryDto.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,5 +11,5 @@ public class RequestPassengerCategoryDto {
 
     private final String categoryName;
 
-    private final BigDecimal discountPercentage;
+    private final float discountPercentage;
 }
